each screen draws own DOM, instead of sharing table.
some of the screens were getting complex ("next" buttons you click,
previous buttons, input categories vs. trial categories, the "X"
was in the same div as the instructions, with the styles switching
between screens).  this way they're more independent.  there could
be a performance issue? and we need to find a way to keep the trial
categories from flickering.