changes used to output all.csv
this was all the in-lab formatted trials of the IAT.

we filtered out IDs that don't have a "01" or "02" in them. this removed
fake accounts like "timtesting" and "test-abc-123", but left in mistyped
IDs such as "*101*" and "101f".

i also had to flatten out the output-lab-iat-format return value so it
was one long list, instead of lists of lists.