include date when calculating IAT scores
make-readable no longer does it's own looping, it must be called with a
single instance. so anything that calls it must do the `map`ing.

i did this because i didn't want to pass the timestamp through a bunch
of the scoring functions (better not to mess with them unless needed).