output iat1.csv - iat7.csv
calling (write-file-to-iat 0) outputs to iat1.csv and includes
everyone's first iat.

calling (write-file-to-iat 1) outputs to iat2.csv and includes
everyone's second iat.

it pays no mind to how many total IATs a person took, only that they
took n in chronological order. so if a person completed the iat less
5 times, then they won't appear in iat5.csv.